Split claim validation out from id token validation
